refactors api routes to routes + controllers
