+++ routes/index.js
@@ -1,3 +1,4 @@
 module.exports = {
   //put some routes in here
+  auth: require('./auth'),
 };
